chore: fix all 63 clippy warnings across the workspace
Resolve all clippy warnings without changing any behavior:

- Add type aliases (AsyncBranch<T>, DurableFuture<T>, TaskFn<T>) to
  eliminate 27 "very complex type" warnings in test code
- Replace manual div_ceil with .div_ceil() (stable since Rust 1.73)
- Replace manual range checks with .contains()
- Use .is_some_and() instead of .map_or(false, ...)
- Use .prop_map(Some) instead of .prop_map(|s| Some(s))
- Remove redundant casts (u64 as u64, i32 as i32)
- Remove needless borrows and clone on Copy types
- Use if let instead of match for single-pattern destructuring
- Use array literals instead of vec![] for single elements
- Add Default impl for MockDurableServiceClient
- Use struct update syntax instead of field assignment after default
- Replace assert!(true, ...) with meaningful code
- Use !is_empty() instead of len() >= 1

+++ aws-durable-execution-sdk/sdk/tests/replay_scenarios.rs
@@ -142,7 +142,7 @@ async fn test_replay_preserves_result_order() {
         assert!(result.is_ok());
         match result.unwrap() {
             ReplayResult::Replayed(value) => {
-                assert_eq!(value, i as i32, "Step {} should return {}", i, i)
+                assert_eq!(value, i, "Step {} should return {}", i, i)
             }
             _ => panic!("Expected Replayed result for step-{}", i),
         }
